Remove description snippet from WatchLaterVideos component
