+++ client/src/components/Header/Header.jsx
@@ -1,20 +1,150 @@
 import React from "react";
 import { Link, useNavigate } from "react-router";
 import SearchIcon from "../SearchIcon";
-import { useDispatch, useSelector } from "react-redux";
+import { useSelector } from "react-redux";
 import { Button } from "../ui/button";
-import { useLogout } from "../../hooks/useLogout.hook";
-import { setEdit } from "../../slice/editSlice";
 import { useState } from "react";
 import { LoginDialog } from "../Dialogs/LoginAlertDialog/LoginAlertDialog";
 import SidebarMobile from "../SideBarMobile/SidebarMobile";
+import {
+    Clock,
+    FolderOpen,
+    HelpCircle,
+    History,
+    Home,
+    Settings,
+    ThumbsUp,
+    Users,
+    Menu
+} from "lucide-react";
+import {
+    Sheet,
+    SheetClose,
+    SheetContent,
+    SheetHeader,
+    SheetTitle,
+    SheetTrigger,
+    SheetDescription
+} from "@/components/ui/sheet";
+
+const navigationItems = [
+    {
+        id: "home",
+        label: "Home",
+        icon: Home,
+        route: "/",
+        requiresAuth: false,
+        showOnMobile: true
+    },
+    {
+        id: "liked",
+        label: "Liked Videos",
+        icon: ThumbsUp,
+        route: "/liked-videos",
+        requiresAuth: true,
+        showOnMobile: false
+    },
+    {
+        id: "history",
+        label: "History",
+        icon: History,
+        route: "/watch-history",
+        requiresAuth: true,
+        showOnMobile: true
+    },
+    {
+        id: "watchLater",
+        label: "Watch Later",
+        icon: Clock,
+        route: "/watch-later",
+        requiresAuth: true,
+        showOnMobile: false
+    },
+    {
+        id: "playlists",
+        label: "My Playlists",
+        icon: FolderOpen,
+        route: "/my-playlists",
+        requiresAuth: true,
+        showOnMobile: true
+    },
+    {
+        id: "subscribers",
+        label: "Subscribers",
+        icon: Users,
+        route: "/subscribers",
+        requiresAuth: true,
+        showOnMobile: true
+    }
+];
+
+const bottomNavigationItems = [
+    {
+        id: "support",
+        label: "Support",
+        icon: HelpCircle,
+        route: "/support",
+        requiresAuth: false
+    },
+    {
+        id: "settings",
+        label: "Settings",
+        icon: Settings,
+        route: "/settings",
+        requiresAuth: false
+    }
+];
 
 function Header() {
-    const navigator = useNavigate();
     const isLoggedIn = useSelector((state) => state.loginUser.login_user.user);
-    const logout = useLogout();
-    const dispatch = useDispatch();
+    // const logout = useLogout();
     const [loginDialogOpen, setLoginDialogOpen] = useState(false);
+    const navigate = useNavigate();
+    const [sheetOpen, setSheetOpen] = useState(false);
+
+    const handleNavigation = (item) => {
+        if (!isLoggedIn) {
+            setLoginDialogOpen(true);
+            setSheetOpen(false);
+            return;
+        }
+        navigate(item.route);
+        setSheetOpen(false);
+    };
+
+    const NavItem = ({ item, isMobile = false }) => {
+        const IconComponent = item.icon;
+
+        return (
+            <button
+                onClick={() => handleNavigation(item)}
+                className={`
+                        flex items-center w-full px-4 py-3 text-left transition-all duration-200
+                        ${
+                            isMobile
+                                ? "hover:bg-neutral-800 rounded-lg text-white hover:text-primary"
+                                : "flex-col sm:flex-row sm:px-3 sm:py-2 border border-transparent hover:bg-primary hover:text-black focus:bg-primary focus:text-black rounded-lg"
+                        }
+                    `}
+                aria-label={item.label}
+            >
+                <IconComponent
+                    className={`
+                            w-5 h-5 flex-shrink-0
+                            ${isMobile ? "mr-3" : "sm:mr-3 lg:mr-3"}
+                        `}
+                />
+                <span
+                    className={`
+                        text-sm font-medium
+                        ${isMobile ? "block" : "hidden sm:block"}
+                    `}
+                >
+                    {item.label}
+                </span>
+            </button>
+        );
+    };
 
     return (
         <header className="sticky inset-x-0 top-0 z-50 w-full border-b border-white bg-[#121212] px-4">
@@ -57,217 +187,120 @@ function Header() {
                     </svg>
                 </button>
                 <button className="group peer ml-4 flex w-6 shrink-0 flex-wrap gap-y-1.5 sm:hidden">
-                    <SidebarMobile />   
-                </button>
-                <div className="fixed inset-y-0 right-0 flex w-full max-w-xs shrink-0 translate-x-full flex-col border-l border-white bg-[#121212] duration-200 hover:translate-x-0 peer-focus:translate-x-0 sm:static sm:ml-4 sm:w-auto sm:translate-x-0 sm:border-none">
-                    <div className="relative flex w-full items-center justify-between border-b border-white px-4 py-2 sm:hidden">
-                        <span className="inline-block w-12">
-                            <img src="/smallLogo.png" alt="logo" className="" />
-                        </span>
-                        <button className="inline-block w-8">
-                            {/* <svg
-                                xmlns="http://www.w3.org/2000/svg"
-                                fill="none"
-                                viewBox="0 0 24 24"
-                                strokeWidth="1.5"
-                                stroke="currentColor"
-                                aria-hidden="true"
-                            >
-                                <path
-                                    strokeLinecap="round"
-                                    strokeLinejoin="round"
-                                    d="M9.75 9.75l4.5 4.5m0-4.5l-4.5 4.5M21 12a9 9 0 11-18 0 9 9 0 0118 0z"
-                                ></path>
-                            </svg> */}
-                        </button>
-                    </div>
-                    <ul className="my-4 flex w-full flex-wrap gap-2 px-4 sm:hidden">
-                        <li className="w-full">
-                            <button
-                                className="flex w-full items-center justify-start gap-x-4 border border-white px-4 py-1.5 text-left hover:bg-[#ae7aff] hover:text-black focus:border-[#ae7aff] focus:bg-[#ae7aff] focus:text-black"
-                                onClick={() => {
-                                    if (
-                                        isLoggedIn === null ||
-                                        isLoggedIn === undefined
-                                    ) {
-                                        setLoginDialogOpen(true);
-                                        return;
-                                    }
-                                    navigator("/liked-videos");
-                                }}
-                            >
-                                <span className="inline-block w-full max-w-[20px] group-hover:mr-4 lg:mr-4">
-                                    <svg
-                                        style={{ width: "100%" }}
-                                        viewBox="0 0 22 22"
-                                        fill="none"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                    >
-                                        <path
-                                            d="M6 21V10M1 12V19C1 20.1046 1.89543 21 3 21H16.4262C17.907 21 19.1662 19.9197 19.3914 18.4562L20.4683 11.4562C20.7479 9.6389 19.3418 8 17.5032 8H14C13.4477 8 13 7.55228 13 7V3.46584C13 2.10399 11.896 1 10.5342 1C10.2093 1 9.91498 1.1913 9.78306 1.48812L6.26394 9.40614C6.10344 9.76727 5.74532 10 5.35013 10H3C1.89543 10 1 10.8954 1 12Z"
-                                            stroke="currentColor"
-                                            strokeWidth="2"
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                        ></path>
-                                    </svg>
-                                </span>
-                                <span>Liked Videos</span>
-                            </button>
-                        </li>
-                        <li className="w-full">
-                            <button
-                                onClick={() => {
-                                    if (
-                                        isLoggedIn === null ||
-                                        isLoggedIn === undefined
-                                    ) {
-                                        setLoginDialogOpen(true);
-                                        return;
-                                    }
-                                    navigator("/watch-later");
-                                }}
-                                className="flex w-full items-center justify-start gap-x-4 border border-white px-4 py-1.5 text-left hover:bg-[#ae7aff] hover:text-black focus:border-[#ae7aff] focus:bg-[#ae7aff] focus:text-black"
-                            >
-                                <span className="inline-block w-full max-w-[20px] group-hover:mr-4 lg:mr-4">
-                                    <svg
-                                        style={{ width: "100%" }}
-                                        viewBox="0 0 22 16"
-                                        fill="none"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                    >
-                                        <path
-                                            d="M21 4.93137C21 4.32555 21 4.02265 20.8802 3.88238C20.7763 3.76068 20.6203 3.69609 20.4608 3.70865C20.2769 3.72312 20.0627 3.93731 19.6343 4.36569L16 8L19.6343 11.6343C20.0627 12.0627 20.2769 12.2769 20.4608 12.2914C20.6203 12.3039 20.7763 12.2393 20.8802 12.1176C21 11.9774 21 11.6744 21 11.0686V4.93137Z"
-                                            stroke="currentColor"
-                                            strokeWidth="2"
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                        ></path>
-                                        <path
-                                            d="M1 5.8C1 4.11984 1 3.27976 1.32698 2.63803C1.6146 2.07354 2.07354 1.6146 2.63803 1.32698C3.27976 1 4.11984 1 5.8 1H11.2C12.8802 1 13.7202 1 14.362 1.32698C14.9265 1.6146 15.3854 2.07354 15.673 2.63803C16 3.27976 16 4.11984 16 5.8V10.2C16 11.8802 16 12.7202 15.673 13.362C15.3854 13.9265 14.9265 14.3854 14.362 14.673C13.7202 15 12.8802 15 11.2 15H5.8C4.11984 15 3.27976 15 2.63803 14.673C2.07354 14.3854 1.6146 13.9265 1.32698 13.362C1 12.7202 1 11.8802 1 10.2V5.8Z"
-                                            stroke="currentColor"
-                                            strokeWidth="2"
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                        ></path>
-                                    </svg>
-                                </span>
-                                <span>Watch Later</span>
-                            </button>
-                        </li>
-                        <li className="w-full">
-                            <button className="flex w-full items-center justify-start gap-x-4 border border-white px-4 py-1.5 text-left hover:bg-[#ae7aff] hover:text-black focus:border-[#ae7aff] focus:bg-[#ae7aff] focus:text-black">
-                                <span className="inline-block w-full max-w-[20px] group-hover:mr-4 lg:mr-4">
-                                    <svg
-                                        style={{ width: "100%" }}
-                                        viewBox="0 0 22 22"
-                                        fill="none"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                    >
-                                        <path
-                                            d="M8.09 8C8.3251 7.33167 8.78915 6.76811 9.39995 6.40913C10.0108 6.05016 10.7289 5.91894 11.4272 6.03871C12.1255 6.15849 12.7588 6.52152 13.2151 7.06353C13.6713 7.60553 13.9211 8.29152 13.92 9C13.92 11 10.92 12 10.92 12M11 16H11.01M21 11C21 16.5228 16.5228 21 11 21C5.47715 21 1 16.5228 1 11C1 5.47715 5.47715 1 11 1C16.5228 1 21 5.47715 21 11Z"
-                                            stroke="currentColor"
-                                            strokeWidth="2"
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                        ></path>
-                                    </svg>
-                                </span>
-                                <span>Support</span>
-                            </button>
-                        </li>
-                        <li className="w-full">
-                            <button className="flex w-full items-center justify-start gap-x-4 border border-white px-4 py-1.5 text-left hover:bg-[#ae7aff] hover:text-black focus:border-[#ae7aff] focus:bg-[#ae7aff] focus:text-black">
-                                <span className="inline-block w-full max-w-[20px] group-hover:mr-4 lg:mr-4">
-                                    <svg
-                                        style={{ width: "100%" }}
-                                        viewBox="0 0 22 22"
-                                        fill="none"
-                                        xmlns="http://www.w3.org/2000/svg"
-                                    >
-                                        <path
-                                            d="M11 14C12.6569 14 14 12.6569 14 11C14 9.34315 12.6569 8 11 8C9.34315 8 8 9.34315 8 11C8 12.6569 9.34315 14 11 14Z"
-                                            stroke="currentColor"
-                                            strokeWidth="2"
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                        ></path>
-                                        <path
-                                            d="M17.7273 13.7273C17.6063 14.0015 17.5702 14.3056 17.6236 14.6005C17.6771 14.8954 17.8177 15.1676 18.0273 15.3818L18.0818 15.4364C18.2509 15.6052 18.385 15.8057 18.4765 16.0265C18.568 16.2472 18.6151 16.4838 18.6151 16.7227C18.6151 16.9617 18.568 17.1983 18.4765 17.419C18.385 17.6397 18.2509 17.8402 18.0818 18.0091C17.913 18.1781 17.7124 18.3122 17.4917 18.4037C17.271 18.4952 17.0344 18.5423 16.7955 18.5423C16.5565 18.5423 16.3199 18.4952 16.0992 18.4037C15.8785 18.3122 15.678 18.1781 15.5091 18.0091L15.4545 17.9545C15.2403 17.745 14.9682 17.6044 14.6733 17.5509C14.3784 17.4974 14.0742 17.5335 13.8 17.6545C13.5311 17.7698 13.3018 17.9611 13.1403 18.205C12.9788 18.4489 12.8921 18.7347 12.8909 19.0273V19.1818C12.8909 19.664 12.6994 20.1265 12.3584 20.4675C12.0174 20.8084 11.5549 21 11.0727 21C10.5905 21 10.1281 20.8084 9.78708 20.4675C9.4461 20.1265 9.25455 19.664 9.25455 19.1818V19.1C9.24751 18.7991 9.15011 18.5073 8.97501 18.2625C8.79991 18.0176 8.55521 17.8312 8.27273 17.7273C7.99853 17.6063 7.69437 17.5702 7.39947 17.6236C7.10456 17.6771 6.83244 17.8177 6.61818 18.0273L6.56364 18.0818C6.39478 18.2509 6.19425 18.385 5.97353 18.4765C5.7528 18.568 5.51621 18.6151 5.27727 18.6151C5.03834 18.6151 4.80174 18.568 4.58102 18.4765C4.36029 18.385 4.15977 18.2509 3.99091 18.0818C3.82186 17.913 3.68775 17.7124 3.59626 17.4917C3.50476 17.271 3.45766 17.0344 3.45766 16.7955C3.45766 16.5565 3.50476 16.3199 3.59626 16.0992C3.68775 15.8785 3.82186 15.678 3.99091 15.5091L4.04545 15.4545C4.25503 15.2403 4.39562 14.9682 4.4491 14.6733C4.50257 14.3784 4.46647 14.0742 4.34545 13.8C4.23022 13.5311 4.03887 13.3018 3.79497 13.1403C3.55107 12.9788 3.26526 12.8921 2.97273 12.8909H2.81818C2.33597 12.8909 1.87351 12.6994 1.53253 12.3584C1.19156 12.0174 1 11.5549 1 11.0727C1 10.5905 1.19156 10.1281 1.53253 9.78708C1.87351 9.4461 2.33597 9.25455 2.81818 9.25455H2.9C3.2009 9.24751 3.49273 9.15011 3.73754 8.97501C3.98236 8.79991 4.16883 8.55521 4.27273 8.27273C4.39374 7.99853 4.42984 7.69437 4.37637 7.39947C4.3229 7.10456 4.18231 6.83244 3.97273 6.61818L3.91818 6.56364C3.74913 6.39478 3.61503 6.19425 3.52353 5.97353C3.43203 5.7528 3.38493 5.51621 3.38493 5.27727C3.38493 5.03834 3.43203 4.80174 3.52353 4.58102C3.61503 4.36029 3.74913 4.15977 3.91818 3.99091C4.08704 3.82186 4.28757 3.68775 4.50829 3.59626C4.72901 3.50476 4.96561 3.45766 5.20455 3.45766C5.44348 3.45766 5.68008 3.50476 5.9008 3.59626C6.12152 3.68775 6.32205 3.82186 6.49091 3.99091L6.54545 4.04545C6.75971 4.25503 7.03183 4.39562 7.32674 4.4491C7.62164 4.50257 7.9258 4.46647 8.2 4.34545H8.27273C8.54161 4.23022 8.77093 4.03887 8.93245 3.79497C9.09397 3.55107 9.18065 3.26526 9.18182 2.97273V2.81818C9.18182 2.33597 9.37338 1.87351 9.71435 1.53253C10.0553 1.19156 10.5178 1 11 1C11.4822 1 11.9447 1.19156 12.2856 1.53253C12.6266 1.87351 12.8182 2.33597 12.8182 2.81818V2.9C12.8193 3.19253 12.906 3.47834 13.0676 3.72224C13.2291 3.96614 13.4584 4.15749 13.7273 4.27273C14.0015 4.39374 14.3056 4.42984 14.6005 4.37637C14.8954 4.3229 15.1676 4.18231 15.3818 3.97273L15.4364 3.91818C15.6052 3.74913 15.8057 3.61503 16.0265 3.52353C16.2472 3.43203 16.4838 3.38493 16.7227 3.38493C16.9617 3.38493 17.1983 3.43203 17.419 3.52353C17.6397 3.61503 17.8402 3.74913 18.0091 3.91818C18.1781 4.08704 18.3122 4.28757 18.4037 4.50829C18.4952 4.72901 18.5423 4.96561 18.5423 5.20455C18.5423 5.44348 18.4952 5.68008 18.4037 5.9008C18.3122 6.12152 18.1781 6.32205 18.0091 6.49091L17.9545 6.54545C17.745 6.75971 17.6044 7.03183 17.5509 7.32674C17.4974 7.62164 17.5335 7.9258 17.6545 8.2V8.27273C17.7698 8.54161 17.9611 8.77093 18.205 8.93245C18.4489 9.09397 18.7347 9.18065 19.0273 9.18182H19.1818C19.664 9.18182 20.1265 9.37338 20.4675 9.71435C20.8084 10.0553 21 10.5178 21 11C21 11.4822 20.8084 11.9447 20.4675 12.2856C20.1265 12.6266 19.664 12.8182 19.1818 12.8182H19.1C18.8075 12.8193 18.5217 12.906 18.2778 13.0676C18.0339 13.2291 17.8425 13.4584 17.7273 13.7273Z"
-                                            stroke="currentColor"
-                                            strokeWidth="2"
-                                            strokeLinecap="round"
-                                            strokeLinejoin="round"
-                                        ></path>
-                                    </svg>
-                                </span>
-                                <span>Settings</span>
-                            </button>
-                        </li>
-                    </ul>
-                    {isLoggedIn ? (
-                        <div
-                            className="flex items-center justify-between gap-6 px-4 py-2 
-  bg-background border border-border rounded-xl shadow-sm"
+                    <Sheet open={sheetOpen} onOpenChange={setSheetOpen}>
+                        <SheetTrigger
+                            className="scale-135 mt-1.5 mr-1.5 "
+                            asChild
                         >
-                            {/* Avatar Link */}
-                            <Link
-                                to={`/channel/${isLoggedIn.username}/videos`}
-                                onClick={() => {
-                                    dispatch(setEdit(false));
-                                }}
-                                className="relative group"
-                            >
-                                <img
-                                    src={isLoggedIn.avatar.url}
-                                    alt="avatar"
-                                    height={48}
-                                    width={48}
-                                    onError={(e) =>
-                                        (e.currentTarget.src = "/user.png")
-                                    }
-                                    className="h-12 w-12 rounded-full border-2 border-primary 
-        transition duration-300 group-hover:scale-105 group-hover:shadow-md"
-                                />
-                                {/* Online Status */}
-                                <span
-                                    className="absolute -bottom-1 -right-1 h-3 w-3 rounded-full 
-      bg-green-500 border border-background"
-                                ></span>
-                            </Link>
+                            <button className="flex flex-col items-center justify-center p-2 text-white hover:text-white transition-colors duration-200">
+                                <Menu className="w-5 h-5 mb-1" />
+                            </button>
+                        </SheetTrigger>
 
-                            {/* Actions */}
-                            <div>
-                                <Button
-                                    variant={"destructive"}
-                                    onClick={logout}
-                                    className="cursor-pointer"
-                                >
-                                    Logout
-                                </Button>
+                        <SheetContent
+                            side="right"
+                            className="w-80 bg-[#121212] border-l border-neutral-800 p-0"
+                        >
+                            <SheetHeader className="px-6 py-4 border-b border-neutral-800">
+                                <div className="flex items-center justify-between">
+                                    <SheetTitle className="text-white text-lg font-semibold">
+                                        Menu
+                                    </SheetTitle>
+                                </div>
+                                <SheetDescription className="text-neutral-400 text-sm">
+                                    Access all your favorite features
+                                </SheetDescription>
+                            </SheetHeader>
+
+                            <div className="flex flex-col h-full">
+                                {/* User Section */}
+                                {isLoggedIn && (
+                                    <div className="px-6 py-4 border-b border-neutral-800">
+                                        <div className="flex items-center space-x-3">
+                                            <div className="w-10 h-10 bg-primary rounded-full flex items-center justify-center">
+                                                <span className="text-black font-semibold text-sm">
+                                                    {isLoggedIn?.fullName?.charAt(
+                                                        0
+                                                    ) || "U"}
+                                                </span>
+                                            </div>
+                                            <div>
+                                                <p className="text-white font-medium">
+                                                    {isLoggedIn?.fullName ||
+                                                        "User"}
+                                                </p>
+                                                <p className="text-neutral-400 text-sm">
+                                                    {isLoggedIn?.email ||
+                                                        "user@example.com"}
+                                                </p>
+                                            </div>
+                                        </div>
+                                    </div>
+                                )}
+
+                                {/* Main Navigation */}
+                                <nav className="flex-1 px-2 py-4 space-y-1">
+                                    <div className="px-4 py-2">
+                                        <h3 className="text-neutral-400 text-xs font-semibold uppercase tracking-wide">
+                                            Library
+                                        </h3>
+                                    </div>
+                                    {navigationItems.map((item) => (
+                                        <NavItem
+                                            key={item.id}
+                                            item={item}
+                                            isMobile={true}
+                                        />
+                                    ))}
+                                </nav>
+
+                                {/* Bottom Section */}
+                                <div className="px-2 py-4 border-t border-neutral-800 space-y-1">
+                                    <div className="px-4 py-2">
+                                        <h3 className="text-neutral-400 text-xs font-semibold uppercase tracking-wide">
+                                            More
+                                        </h3>
+                                    </div>
+                                    {bottomNavigationItems.map((item) => (
+                                        <NavItem
+                                            key={item.id}
+                                            item={item}
+                                            isMobile={true}
+                                        />
+                                    ))}
+
+                                    {/* Sign Out Button */}
+                                    {isLoggedIn && (
+                                        <button
+                                            onClick={() => {
+                                                // Add your sign out logic here
+                                                setSheetOpen(false);
+                                            }}
+                                            className="flex items-center w-full px-4 py-3 text-left text-red-400 hover:bg-red-500/10 rounded-lg transition-colors duration-200"
+                                        >
+                                            <svg
+                                                className="w-5 h-5 mr-3"
+                                                fill="none"
+                                                stroke="currentColor"
+                                                viewBox="0 0 24 24"
+                                            >
+                                                <path
+                                                    strokeLinecap="round"
+                                                    strokeLinejoin="round"
+                                                    strokeWidth={2}
+                                                    d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"
+                                                />
+                                            </svg>
+                                            <span className="text-sm font-medium">
+                                                Sign Out
+                                            </span>
+                                        </button>
+                                    )}
+                                </div>
                             </div>
-                        </div>
-                    ) : (
-                        <div className="mb-8 mt-auto flex w-full flex-wrap gap-4 px-4 sm:mb-0 sm:mt-0 sm:items-center sm:px-0">
-                            <button
-                                onClick={() => {
-                                    navigator("/login");
-                                }}
-                                className="w-full bg-[#383737] px-3 py-2 hover:bg-[#4f4e4e] sm:w-auto sm:bg-transparent"
-                            >
-                                Log in
-                            </button>
-                            <button
-                                onClick={() => {
-                                    navigator("/register");
-                                }}
-                                className="mr-1 w-full bg-[#ae7aff] px-3 py-2 text-center font-bold text-black shadow-[5px_5px_0px_0px_#4f4e4e] transition-all duration-150 ease-in-out active:translate-x-[5px] active:translate-y-[5px] active:shadow-[0px_0px_0px_0px_#4f4e4e] sm:w-auto"
-                            >
-                                Sign up
-                            </button>
-                        </div>
-                    )}
-                </div>
+                        </SheetContent>
+                    </Sheet>
+                </button>
             </nav>
         </header>
     );
