+++ client/src/components/SideBarMobile/SidebarMobile.jsx
@@ -342,39 +342,7 @@ function SidebarMobile({ className }) {
         setSheetOpen(false);
     };
 
-    const NavItem = ({ item, isMobile = false }) => {
-        const IconComponent = item.icon;
-
-        return (
-            <button
-                onClick={() => handleNavigation(item)}
-                className={`
-                    flex items-center w-full px-4 py-3 text-left transition-all duration-200
-                    ${
-                        isMobile
-                            ? "hover:bg-neutral-800 rounded-lg text-white hover:text-primary"
-                            : "flex-col sm:flex-row sm:px-3 sm:py-2 border border-transparent hover:bg-primary hover:text-black focus:bg-primary focus:text-black rounded-lg"
-                    }
-                `}
-                aria-label={item.label}
-            >
-                <IconComponent
-                    className={`
-                        w-5 h-5 flex-shrink-0
-                        ${isMobile ? "mr-3" : "sm:mr-3 lg:mr-3"}
-                    `}
-                />
-                <span
-                    className={`
-                    text-sm font-medium
-                    ${isMobile ? "block" : "hidden sm:block"}
-                `}
-                >
-                    {item.label}
-                </span>
-            </button>
-        );
-    };
+   
 
     return (
         <>
@@ -382,32 +350,6 @@ function SidebarMobile({ className }) {
                 open={loginDialogOpen}
                 onClose={() => setLoginDialogOpen(false)}
             />
-
-            {/* Desktop Sidebar */}
-            <aside
-                className={`
-                hidden sm:block fixed left-0 top-0 z-40 h-screen w-16 lg:w-64
-                bg-[#121212] border-r border-neutral-800 transition-all duration-300
-                hover:w-64 lg:hover:w-64 overflow-hidden
-                ${className}
-            `}
-            >
-                <div className="flex flex-col h-full pt-20">
-                    {/* Main Navigation */}
-                    <nav className="flex-1 px-2 py-4 space-y-2">
-                        {navigationItems.map((item) => (
-                            <NavItem key={item.id} item={item} />
-                        ))}
-                    </nav>
-
-                    {/* Bottom Navigation */}
-                    <nav className="px-2 py-4 border-t border-neutral-800 space-y-2">
-                        {bottomNavigationItems.map((item) => (
-                            <NavItem key={item.id} item={item} />
-                        ))}
-                    </nav>
-                </div>
-            </aside>
 
             {/* Mobile Navigation */}
             <div className="sm:hidden">
@@ -433,123 +375,7 @@ function SidebarMobile({ className }) {
                                     </button>
                                 );
                             })}
-
-                        {/* Menu Button */}
-                        <Sheet open={sheetOpen} onOpenChange={setSheetOpen}>
-                            <SheetTrigger className="bg-amber-300" asChild>
-                                <button
-                                    className="flex flex-col items-center justify-center p-2 text-neutral-400 hover:text-white transition-colors duration-200"
-                                >
-                                    <Menu className="w-5 h-5 mb-1" />
-                                    <span className="text-xs font-medium">
-                                        More
-                                    </span>
-                                </button>
-                            </SheetTrigger>
-
-                            <SheetContent
-                                side="right"
-                                className="w-80 bg-[#121212] border-l border-neutral-800 p-0"
-                            >
-                                <SheetHeader className="px-6 py-4 border-b border-neutral-800">
-                                    <div className="flex items-center justify-between">
-                                        <SheetTitle className="text-white text-lg font-semibold">
-                                            Menu
-                                        </SheetTitle>
-                                    </div>
-                                    <SheetDescription className="text-neutral-400 text-sm">
-                                        Access all your favorite features
-                                    </SheetDescription>
-                                </SheetHeader>
-
-                                <div className="flex flex-col h-full">
-                                    {/* User Section */}
-                                    {isLoggedIn && (
-                                        <div className="px-6 py-4 border-b border-neutral-800">
-                                            <div className="flex items-center space-x-3">
-                                                <div className="w-10 h-10 bg-primary rounded-full flex items-center justify-center">
-                                                    <span className="text-black font-semibold text-sm">
-                                                        {isLoggedIn?.fullName?.charAt(
-                                                            0
-                                                        ) || "U"}
-                                                    </span>
-                                                </div>
-                                                <div>
-                                                    <p className="text-white font-medium">
-                                                        {isLoggedIn?.fullName ||
-                                                            "User"}
-                                                    </p>
-                                                    <p className="text-neutral-400 text-sm">
-                                                        {isLoggedIn?.email ||
-                                                            "user@example.com"}
-                                                    </p>
-                                                </div>
-                                            </div>
-                                        </div>
-                                    )}
-
-                                    {/* Main Navigation */}
-                                    <nav className="flex-1 px-2 py-4 space-y-1">
-                                        <div className="px-4 py-2">
-                                            <h3 className="text-neutral-400 text-xs font-semibold uppercase tracking-wide">
-                                                Library
-                                            </h3>
-                                        </div>
-                                        {navigationItems.map((item) => (
-                                            <NavItem
-                                                key={item.id}
-                                                item={item}
-                                                isMobile={true}
-                                            />
-                                        ))}
-                                    </nav>
-
-                                    {/* Bottom Section */}
-                                    <div className="px-2 py-4 border-t border-neutral-800 space-y-1">
-                                        <div className="px-4 py-2">
-                                            <h3 className="text-neutral-400 text-xs font-semibold uppercase tracking-wide">
-                                                More
-                                            </h3>
-                                        </div>
-                                        {bottomNavigationItems.map((item) => (
-                                            <NavItem
-                                                key={item.id}
-                                                item={item}
-                                                isMobile={true}
-                                            />
-                                        ))}
-
-                                        {/* Sign Out Button */}
-                                        {isLoggedIn && (
-                                            <button
-                                                onClick={() => {
-                                                    // Add your sign out logic here
-                                                    setSheetOpen(false);
-                                                }}
-                                                className="flex items-center w-full px-4 py-3 text-left text-red-400 hover:bg-red-500/10 rounded-lg transition-colors duration-200"
-                                            >
-                                                <svg
-                                                    className="w-5 h-5 mr-3"
-                                                    fill="none"
-                                                    stroke="currentColor"
-                                                    viewBox="0 0 24 24"
-                                                >
-                                                    <path
-                                                        strokeLinecap="round"
-                                                        strokeLinejoin="round"
-                                                        strokeWidth={2}
-                                                        d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1"
-                                                    />
-                                                </svg>
-                                                <span className="text-sm font-medium">
-                                                    Sign Out
-                                                </span>
-                                            </button>
-                                        )}
-                                    </div>
-                                </div>
-                            </SheetContent>
-                        </Sheet>
+                       
                     </div>
                 </nav>
