+++ client/src/components/Sidebar/Sidebar.jsx
@@ -15,417 +15,273 @@ import { useSelector } from "react-redux";
 import { useState } from "react";
 import { LoginDialog } from "../Dialogs/LoginAlertDialog/LoginAlertDialog";
 
-// function Sidebar({ className }) {
-//     const navigator = useNavigate();
-//     const [loginDialogOpen, setLoginDialogOpen] = useState(false);
-//     const isLoggedIn = useSelector((state) => state.loginUser.login_user.user);
-
-//     return (
-//         <aside
-//             className={`group fixed inset-x-0 bottom-0 z-40 w-full shrink-0 border-t border-white bg-[#121212] px-2 py-2 sm:absolute sm:inset-y-0 sm:max-w-[70px] sm:border-r sm:border-t-0 sm:py-6 sm:hover:max-w-[250px] lg:sticky lg:max-w-[250px] ${className}`}
-//         >
-//             <LoginDialog
-//                 open={loginDialogOpen}
-//                 onClose={() => setLoginDialogOpen(false)}
-//             />
-//             <ul className="flex justify-around gap-y-2 sm:sticky sm:top-[106px] sm:min-h-[calc(100vh-130px)] sm:flex-col">
-//                 <li className="">
-//                     <button
-//                         onClick={() => {
-//                             navigator("/");
-//                         }}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                         variant={"default"}
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 20 21"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M6 15.9997H14M9.0177 1.76375L2.23539 7.03888C1.78202 7.3915 1.55534 7.56781 1.39203 7.78861C1.24737 7.9842 1.1396 8.20454 1.07403 8.43881C1 8.70327 1 8.99045 1 9.56481V16.7997C1 17.9198 1 18.4799 1.21799 18.9077C1.40973 19.284 1.71569 19.59 2.09202 19.7818C2.51984 19.9997 3.07989 19.9997 4.2 19.9997H15.8C16.9201 19.9997 17.4802 19.9997 17.908 19.7818C18.2843 19.59 18.5903 19.284 18.782 18.9077C19 18.4799 19 17.9198 19 16.7997V9.56481C19 8.99045 19 8.70327 18.926 8.43881C18.8604 8.20454 18.7526 7.9842 18.608 7.78861C18.4447 7.56781 18.218 7.3915 17.7646 7.03888L10.9823 1.76376C10.631 1.4905 10.4553 1.35388 10.2613 1.30136C10.0902 1.25502 9.9098 1.25502 9.73865 1.30136C9.54468 1.35388 9.36902 1.4905 9.0177 1.76375Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             Home
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="hidden sm:block">
-//                     <button
-//                         onClick={() => {
-//                             if (
-//                                 isLoggedIn === null ||
-//                                 isLoggedIn === undefined
-//                             ) {
-//                                 setLoginDialogOpen(true);
-//                                 return;
-//                             }
-//                             navigator("/liked-videos");
-//                         }}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 22 22"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M6 21V10M1 12V19C1 20.1046 1.89543 21 3 21H16.4262C17.907 21 19.1662 19.9197 19.3914 18.4562L20.4683 11.4562C20.7479 9.6389 19.3418 8 17.5032 8H14C13.4477 8 13 7.55228 13 7V3.46584C13 2.10399 11.896 1 10.5342 1C10.2093 1 9.91498 1.1913 9.78306 1.48812L6.26394 9.40614C6.10344 9.76727 5.74532 10 5.35013 10H3C1.89543 10 1 10.8954 1 12Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             Liked Videos
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="">
-//                     <button
-//                         onClick={() => {
-//                             if (
-//                                 isLoggedIn === null ||
-//                                 isLoggedIn === undefined
-//                             ) {
-//                                 setLoginDialogOpen(true);
-//                                 return;
-//                             }
-//                             navigator("/watch-history");
-//                         }}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 22 20"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M20.7 11.5L18.7005 9.5L16.7 11.5M19 10C19 14.9706 14.9706 19 10 19C5.02944 19 1 14.9706 1 10C1 5.02944 5.02944 1 10 1C13.3019 1 16.1885 2.77814 17.7545 5.42909M10 5V10L13 12"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             History
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="hidden sm:block">
-//                     <button
-//                         onClick={() => {
-//                             if (
-//                                 isLoggedIn === null ||
-//                                 isLoggedIn === undefined
-//                             ) {
-//                                 setLoginDialogOpen(true);
-//                                 return;
-//                             }
-//                             navigator("/watch-later");
-//                         }}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <Clock />
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             Watch Later
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="">
-//                     <button
-//                         onClick={() => {
-//                             if (
-//                                 isLoggedIn === null ||
-//                                 isLoggedIn === undefined
-//                             ) {
-//                                 setLoginDialogOpen(true);
-//                                 return;
-//                             }
-//                             navigator("/my-playlists");
-//                         }}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 22 20"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M12 5L10.8845 2.76892C10.5634 2.1268 10.4029 1.80573 10.1634 1.57116C9.95158 1.36373 9.69632 1.20597 9.41607 1.10931C9.09916 1 8.74021 1 8.02229 1H4.2C3.0799 1 2.51984 1 2.09202 1.21799C1.71569 1.40973 1.40973 1.71569 1.21799 2.09202C1 2.51984 1 3.0799 1 4.2V5M1 5H16.2C17.8802 5 18.7202 5 19.362 5.32698C19.9265 5.6146 20.3854 6.07354 20.673 6.63803C21 7.27976 21 8.11984 21 9.8V14.2C21 15.8802 21 16.7202 20.673 17.362C20.3854 17.9265 19.9265 18.3854 19.362 18.673C18.7202 19 17.8802 19 16.2 19H5.8C4.11984 19 3.27976 19 2.63803 18.673C2.07354 18.3854 1.6146 17.9265 1.32698 17.362C1 16.7202 1 15.8802 1 14.2V5Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             My Playlists
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="">
-//                     <button
-//                         onClick={() => {
-//                             if (
-//                                 isLoggedIn === null ||
-//                                 isLoggedIn === undefined
-//                             ) {
-//                                 setLoginDialogOpen(true);
-//                                 return;
-//                             }
-//                             navigator("/subscribers");
-//                         }}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 22 20"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M11 13.5H6.5C5.10444 13.5 4.40665 13.5 3.83886 13.6722C2.56045 14.06 1.56004 15.0605 1.17224 16.3389C1 16.9067 1 17.6044 1 19M15 16L17 18L21 14M13.5 5.5C13.5 7.98528 11.4853 10 9 10C6.51472 10 4.5 7.98528 4.5 5.5C4.5 3.01472 6.51472 1 9 1C11.4853 1 13.5 3.01472 13.5 5.5Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             Subscribers
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="hidden sm:block mt-auto">
-//                     <button
-//                         onClick={() => navigator("/")}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 22 22"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M8.09 8C8.3251 7.33167 8.78915 6.76811 9.39995 6.40913C10.0108 6.05016 10.7289 5.91894 11.4272 6.03871C12.1255 6.15849 12.7588 6.52152 13.2151 7.06353C13.6713 7.60553 13.9211 8.29152 13.92 9C13.92 11 10.92 12 10.92 12M11 16H11.01M21 11C21 16.5228 16.5228 21 11 21C5.47715 21 1 16.5228 1 11C1 5.47715 5.47715 1 11 1C16.5228 1 21 5.47715 21 11Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             Support
-//                         </span>
-//                     </button>
-//                 </li>
-//                 <li className="hidden sm:block">
-//                     <button
-//                         onClick={() => navigator("/")}
-//                         className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
-//                     >
-//                         <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
-//                             <svg
-//                                 style={{ width: "100%" }}
-//                                 viewBox="0 0 22 22"
-//                                 fill="none"
-//                                 xmlns="http://www.w3.org/2000/svg"
-//                             >
-//                                 <path
-//                                     d="M11 14C12.6569 14 14 12.6569 14 11C14 9.34315 12.6569 8 11 8C9.34315 8 8 9.34315 8 11C8 12.6569 9.34315 14 11 14Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                                 <path
-//                                     d="M17.7273 13.7273C17.6063 14.0015 17.5702 14.3056 17.6236 14.6005C17.6771 14.8954 17.8177 15.1676 18.0273 15.3818L18.0818 15.4364C18.2509 15.6052 18.385 15.8057 18.4765 16.0265C18.568 16.2472 18.6151 16.4838 18.6151 16.7227C18.6151 16.9617 18.568 17.1983 18.4765 17.419C18.385 17.6397 18.2509 17.8402 18.0818 18.0091C17.913 18.1781 17.7124 18.3122 17.4917 18.4037C17.271 18.4952 17.0344 18.5423 16.7955 18.5423C16.5565 18.5423 16.3199 18.4952 16.0992 18.4037C15.8785 18.3122 15.678 18.1781 15.5091 18.0091L15.4545 17.9545C15.2403 17.745 14.9682 17.6044 14.6733 17.5509C14.3784 17.4974 14.0742 17.5335 13.8 17.6545C13.5311 17.7698 13.3018 17.9611 13.1403 18.205C12.9788 18.4489 12.8921 18.7347 12.8909 19.0273V19.1818C12.8909 19.664 12.6994 20.1265 12.3584 20.4675C12.0174 20.8084 11.5549 21 11.0727 21C10.5905 21 10.1281 20.8084 9.78708 20.4675C9.4461 20.1265 9.25455 19.664 9.25455 19.1818V19.1C9.24751 18.7991 9.15011 18.5073 8.97501 18.2625C8.79991 18.0176 8.55521 17.8312 8.27273 17.7273C7.99853 17.6063 7.69437 17.5702 7.39947 17.6236C7.10456 17.6771 6.83244 17.8177 6.61818 18.0273L6.56364 18.0818C6.39478 18.2509 6.19425 18.385 5.97353 18.4765C5.7528 18.568 5.51621 18.6151 5.27727 18.6151C5.03834 18.6151 4.80174 18.568 4.58102 18.4765C4.36029 18.385 4.15977 18.2509 3.99091 18.0818C3.82186 17.913 3.68775 17.7124 3.59626 17.4917C3.50476 17.271 3.45766 17.0344 3.45766 16.7955C3.45766 16.5565 3.50476 16.3199 3.59626 16.0992C3.68775 15.8785 3.82186 15.678 3.99091 15.5091L4.04545 15.4545C4.25503 15.2403 4.39562 14.9682 4.4491 14.6733C4.50257 14.3784 4.46647 14.0742 4.34545 13.8C4.23022 13.5311 4.03887 13.3018 3.79497 13.1403C3.55107 12.9788 3.26526 12.8921 2.97273 12.8909H2.81818C2.33597 12.8909 1.87351 12.6994 1.53253 12.3584C1.19156 12.0174 1 11.5549 1 11.0727C1 10.5905 1.19156 10.1281 1.53253 9.78708C1.87351 9.4461 2.33597 9.25455 2.81818 9.25455H2.9C3.2009 9.24751 3.49273 9.15011 3.73754 8.97501C3.98236 8.79991 4.16883 8.55521 4.27273 8.27273C4.39374 7.99853 4.42984 7.69437 4.37637 7.39947C4.3229 7.10456 4.18231 6.83244 3.97273 6.61818L3.91818 6.56364C3.74913 6.39478 3.61503 6.19425 3.52353 5.97353C3.43203 5.7528 3.38493 5.51621 3.38493 5.27727C3.38493 5.03834 3.43203 4.80174 3.52353 4.58102C3.61503 4.36029 3.74913 4.15977 3.91818 3.99091C4.08704 3.82186 4.28757 3.68775 4.50829 3.59626C4.72901 3.50476 4.96561 3.45766 5.20455 3.45766C5.44348 3.45766 5.68008 3.50476 5.9008 3.59626C6.12152 3.68775 6.32205 3.82186 6.49091 3.99091L6.54545 4.04545C6.75971 4.25503 7.03183 4.39562 7.32674 4.4491C7.62164 4.50257 7.9258 4.46647 8.2 4.34545H8.27273C8.54161 4.23022 8.77093 4.03887 8.93245 3.79497C9.09397 3.55107 9.18065 3.26526 9.18182 2.97273V2.81818C9.18182 2.33597 9.37338 1.87351 9.71435 1.53253C10.0553 1.19156 10.5178 1 11 1C11.4822 1 11.9447 1.19156 12.2856 1.53253C12.6266 1.87351 12.8182 2.33597 12.8182 2.81818V2.9C12.8193 3.19253 12.906 3.47834 13.0676 3.72224C13.2291 3.96614 13.4584 4.15749 13.7273 4.27273C14.0015 4.39374 14.3056 4.42984 14.6005 4.37637C14.8954 4.3229 15.1676 4.18231 15.3818 3.97273L15.4364 3.91818C15.6052 3.74913 15.8057 3.61503 16.0265 3.52353C16.2472 3.43203 16.4838 3.38493 16.7227 3.38493C16.9617 3.38493 17.1983 3.43203 17.419 3.52353C17.6397 3.61503 17.8402 3.74913 18.0091 3.91818C18.1781 4.08704 18.3122 4.28757 18.4037 4.50829C18.4952 4.72901 18.5423 4.96561 18.5423 5.20455C18.5423 5.44348 18.4952 5.68008 18.4037 5.9008C18.3122 6.12152 18.1781 6.32205 18.0091 6.49091L17.9545 6.54545C17.745 6.75971 17.6044 7.03183 17.5509 7.32674C17.4974 7.62164 17.5335 7.9258 17.6545 8.2V8.27273C17.7698 8.54161 17.9611 8.77093 18.205 8.93245C18.4489 9.09397 18.7347 9.18065 19.0273 9.18182H19.1818C19.664 9.18182 20.1265 9.37338 20.4675 9.71435C20.8084 10.0553 21 10.5178 21 11C21 11.4822 20.8084 11.9447 20.4675 12.2856C20.1265 12.6266 19.664 12.8182 19.1818 12.8182H19.1C18.8075 12.8193 18.5217 12.906 18.2778 13.0676C18.0339 13.2291 17.8425 13.4584 17.7273 13.7273Z"
-//                                     stroke="currentColor"
-//                                     strokeWidth="2"
-//                                     strokeLinecap="round"
-//                                     strokeLinejoin="round"
-//                                 ></path>
-//                             </svg>
-//                         </span>
-//                         <span className="block sm:hidden sm:group-hover:inline lg:inline">
-//                             Settings
-//                         </span>
-//                     </button>
-//                 </li>
-//             </ul>
-//         </aside>
-//     );
-// }
 function Sidebar({ className }) {
+    const navigator = useNavigate();
     const [loginDialogOpen, setLoginDialogOpen] = useState(false);
     const isLoggedIn = useSelector((state) => state.loginUser.login_user.user);
 
-    const navigationItems = [
-        {
-            id: "home",
-            label: "Home",
-            icon: Home,
-            route: "/",
-            requiresAuth: false,
-            showOnMobile: true
-        },
-        {
-            id: "liked",
-            label: "Liked Videos",
-            icon: ThumbsUp,
-            route: "/liked-videos",
-            requiresAuth: true,
-            showOnMobile: false
-        },
-        {
-            id: "history",
-            label: "History",
-            icon: History,
-            route: "/watch-history",
-            requiresAuth: true,
-            showOnMobile: true
-        },
-        {
-            id: "watchLater",
-            label: "Watch Later",
-            icon: Clock,
-            route: "/watch-later",
-            requiresAuth: true,
-            showOnMobile: false
-        },
-        {
-            id: "playlists",
-            label: "My Playlists",
-            icon: FolderOpen,
-            route: "/my-playlists",
-            requiresAuth: true,
-            showOnMobile: true
-        },
-        {
-            id: "subscribers",
-            label: "Subscribers",
-            icon: Users,
-            route: "/subscribers",
-            requiresAuth: true,
-            showOnMobile: true
-        }
-    ];
-
-    const bottomNavigationItems = [
-        {
-            id: "support",
-            label: "Support",
-            icon: HelpCircle,
-            route: "/support",
-            requiresAuth: false
-        },
-        {
-            id: "settings",
-            label: "Settings",
-            icon: Settings,
-            route: "/settings",
-            requiresAuth: false
-        }
-    ];
-
-    const NavItem = ({ item, isMobile = false }) => {
-        const IconComponent = item.icon;
-
-        return (
-            <button
-                onClick={() => handleNavigation(item)}
-                className={`
-                    flex items-center w-full px-4 py-3 text-left transition-all duration-200
-                    ${
-                        isMobile
-                            ? "hover:bg-neutral-800 rounded-lg text-white hover:text-primary"
-                            : "flex-col sm:flex-row sm:px-3 sm:py-2 border border-transparent hover:bg-primary hover:text-black focus:bg-primary focus:text-black rounded-lg"
-                    }
-                `}
-                aria-label={item.label}
-            >
-                <IconComponent
-                    className={`
-                        w-5 h-5 flex-shrink-0
-                        ${isMobile ? "mr-3" : "sm:mr-3 lg:mr-3"}
-                    `}
-                />
-                <span
-                    className={`
-                    text-sm font-medium
-                    ${isMobile ? "block" : "hidden sm:block"}
-                `}
-                >
-                    {item.label}
-                </span>
-            </button>
-        );
-    };
-
     return (
-        <>
+        <aside
+            className={`group fixed inset-x-0 bottom-0 z-40 w-full shrink-0 border-t border-white bg-[#121212] px-2 py-2 sm:absolute sm:inset-y-0 sm:max-w-[70px] sm:border-r sm:border-t-0 sm:py-6 sm:hover:max-w-[250px] lg:sticky lg:max-w-[250px] ${className}`}
+        >
             <LoginDialog
                 open={loginDialogOpen}
                 onClose={() => setLoginDialogOpen(false)}
             />
-            <aside
-                className={`
-                hidden sm:block fixed left-0 top-0 z-40 h-screen w-16 lg:w-64
-                bg-[#121212] border-r border-neutral-800 transition-all duration-300
-                hover:w-64 lg:hover:w-64 overflow-hidden
-                ${className}
-            `}
-            >
-                <div className="flex flex-col h-full pt-20">
-                    {/* Main Navigation */}
-                    <nav className="flex-1 px-2 py-4 space-y-2">
-                        {navigationItems.map((item) => (
-                            <NavItem key={item.id} item={item} />
-                        ))}
-                    </nav>
-
-                    {/* Bottom Navigation */}
-                    <nav className="px-2 py-4 border-t border-neutral-800 space-y-2">
-                        {bottomNavigationItems.map((item) => (
-                            <NavItem key={item.id} item={item} />
-                        ))}
-                    </nav>
-                </div>
-            </aside>
-            ;
-        </>
+            <ul className="flex justify-around gap-y-2 sm:sticky sm:top-[106px] sm:min-h-[calc(100vh-130px)] sm:flex-col">
+                <li className="">
+                    <button
+                        onClick={() => {
+                            navigator("/");
+                        }}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                        variant={"default"}
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 20 21"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M6 15.9997H14M9.0177 1.76375L2.23539 7.03888C1.78202 7.3915 1.55534 7.56781 1.39203 7.78861C1.24737 7.9842 1.1396 8.20454 1.07403 8.43881C1 8.70327 1 8.99045 1 9.56481V16.7997C1 17.9198 1 18.4799 1.21799 18.9077C1.40973 19.284 1.71569 19.59 2.09202 19.7818C2.51984 19.9997 3.07989 19.9997 4.2 19.9997H15.8C16.9201 19.9997 17.4802 19.9997 17.908 19.7818C18.2843 19.59 18.5903 19.284 18.782 18.9077C19 18.4799 19 17.9198 19 16.7997V9.56481C19 8.99045 19 8.70327 18.926 8.43881C18.8604 8.20454 18.7526 7.9842 18.608 7.78861C18.4447 7.56781 18.218 7.3915 17.7646 7.03888L10.9823 1.76376C10.631 1.4905 10.4553 1.35388 10.2613 1.30136C10.0902 1.25502 9.9098 1.25502 9.73865 1.30136C9.54468 1.35388 9.36902 1.4905 9.0177 1.76375Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            Home
+                        </span>
+                    </button>
+                </li>
+                <li className="hidden sm:block">
+                    <button
+                        onClick={() => {
+                            if (
+                                isLoggedIn === null ||
+                                isLoggedIn === undefined
+                            ) {
+                                setLoginDialogOpen(true);
+                                return;
+                            }
+                            navigator("/liked-videos");
+                        }}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 22 22"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M6 21V10M1 12V19C1 20.1046 1.89543 21 3 21H16.4262C17.907 21 19.1662 19.9197 19.3914 18.4562L20.4683 11.4562C20.7479 9.6389 19.3418 8 17.5032 8H14C13.4477 8 13 7.55228 13 7V3.46584C13 2.10399 11.896 1 10.5342 1C10.2093 1 9.91498 1.1913 9.78306 1.48812L6.26394 9.40614C6.10344 9.76727 5.74532 10 5.35013 10H3C1.89543 10 1 10.8954 1 12Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            Liked Videos
+                        </span>
+                    </button>
+                </li>
+                <li className="">
+                    <button
+                        onClick={() => {
+                            if (
+                                isLoggedIn === null ||
+                                isLoggedIn === undefined
+                            ) {
+                                setLoginDialogOpen(true);
+                                return;
+                            }
+                            navigator("/watch-history");
+                        }}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 22 20"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M20.7 11.5L18.7005 9.5L16.7 11.5M19 10C19 14.9706 14.9706 19 10 19C5.02944 19 1 14.9706 1 10C1 5.02944 5.02944 1 10 1C13.3019 1 16.1885 2.77814 17.7545 5.42909M10 5V10L13 12"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            History
+                        </span>
+                    </button>
+                </li>
+                <li className="hidden sm:block">
+                    <button
+                        onClick={() => {
+                            if (
+                                isLoggedIn === null ||
+                                isLoggedIn === undefined
+                            ) {
+                                setLoginDialogOpen(true);
+                                return;
+                            }
+                            navigator("/watch-later");
+                        }}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <Clock />
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            Watch Later
+                        </span>
+                    </button>
+                </li>
+                <li className="">
+                    <button
+                        onClick={() => {
+                            if (
+                                isLoggedIn === null ||
+                                isLoggedIn === undefined
+                            ) {
+                                setLoginDialogOpen(true);
+                                return;
+                            }
+                            navigator("/my-playlists");
+                        }}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 22 20"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M12 5L10.8845 2.76892C10.5634 2.1268 10.4029 1.80573 10.1634 1.57116C9.95158 1.36373 9.69632 1.20597 9.41607 1.10931C9.09916 1 8.74021 1 8.02229 1H4.2C3.0799 1 2.51984 1 2.09202 1.21799C1.71569 1.40973 1.40973 1.71569 1.21799 2.09202C1 2.51984 1 3.0799 1 4.2V5M1 5H16.2C17.8802 5 18.7202 5 19.362 5.32698C19.9265 5.6146 20.3854 6.07354 20.673 6.63803C21 7.27976 21 8.11984 21 9.8V14.2C21 15.8802 21 16.7202 20.673 17.362C20.3854 17.9265 19.9265 18.3854 19.362 18.673C18.7202 19 17.8802 19 16.2 19H5.8C4.11984 19 3.27976 19 2.63803 18.673C2.07354 18.3854 1.6146 17.9265 1.32698 17.362C1 16.7202 1 15.8802 1 14.2V5Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            My Playlists
+                        </span>
+                    </button>
+                </li>
+                <li className="">
+                    <button
+                        onClick={() => {
+                            if (
+                                isLoggedIn === null ||
+                                isLoggedIn === undefined
+                            ) {
+                                setLoginDialogOpen(true);
+                                return;
+                            }
+                            navigator("/subscribers");
+                        }}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 22 20"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M11 13.5H6.5C5.10444 13.5 4.40665 13.5 3.83886 13.6722C2.56045 14.06 1.56004 15.0605 1.17224 16.3389C1 16.9067 1 17.6044 1 19M15 16L17 18L21 14M13.5 5.5C13.5 7.98528 11.4853 10 9 10C6.51472 10 4.5 7.98528 4.5 5.5C4.5 3.01472 6.51472 1 9 1C11.4853 1 13.5 3.01472 13.5 5.5Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            Subscribers
+                        </span>
+                    </button>
+                </li>
+                <li className="hidden sm:block mt-auto">
+                    <button
+                        onClick={() => navigator("/")}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 22 22"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M8.09 8C8.3251 7.33167 8.78915 6.76811 9.39995 6.40913C10.0108 6.05016 10.7289 5.91894 11.4272 6.03871C12.1255 6.15849 12.7588 6.52152 13.2151 7.06353C13.6713 7.60553 13.9211 8.29152 13.92 9C13.92 11 10.92 12 10.92 12M11 16H11.01M21 11C21 16.5228 16.5228 21 11 21C5.47715 21 1 16.5228 1 11C1 5.47715 5.47715 1 11 1C16.5228 1 21 5.47715 21 11Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            Support
+                        </span>
+                    </button>
+                </li>
+                <li className="hidden sm:block">
+                    <button
+                        onClick={() => navigator("/")}
+                        className="flex flex-col items-center justify-center border-white py-1 focus:text-[#ae7aff] sm:w-full sm:flex-row sm:border sm:p-1.5 sm:hover:bg-[#ae7aff] sm:hover:text-black sm:focus:border-[#ae7aff] sm:focus:bg-[#ae7aff] sm:focus:text-black sm:group-hover:justify-start sm:group-hover:px-4 lg:justify-start lg:px-4"
+                    >
+                        <span className="inline-block w-5 shrink-0 sm:group-hover:mr-4 lg:mr-4">
+                            <svg
+                                style={{ width: "100%" }}
+                                viewBox="0 0 22 22"
+                                fill="none"
+                                xmlns="http://www.w3.org/2000/svg"
+                            >
+                                <path
+                                    d="M11 14C12.6569 14 14 12.6569 14 11C14 9.34315 12.6569 8 11 8C9.34315 8 8 9.34315 8 11C8 12.6569 9.34315 14 11 14Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                                <path
+                                    d="M17.7273 13.7273C17.6063 14.0015 17.5702 14.3056 17.6236 14.6005C17.6771 14.8954 17.8177 15.1676 18.0273 15.3818L18.0818 15.4364C18.2509 15.6052 18.385 15.8057 18.4765 16.0265C18.568 16.2472 18.6151 16.4838 18.6151 16.7227C18.6151 16.9617 18.568 17.1983 18.4765 17.419C18.385 17.6397 18.2509 17.8402 18.0818 18.0091C17.913 18.1781 17.7124 18.3122 17.4917 18.4037C17.271 18.4952 17.0344 18.5423 16.7955 18.5423C16.5565 18.5423 16.3199 18.4952 16.0992 18.4037C15.8785 18.3122 15.678 18.1781 15.5091 18.0091L15.4545 17.9545C15.2403 17.745 14.9682 17.6044 14.6733 17.5509C14.3784 17.4974 14.0742 17.5335 13.8 17.6545C13.5311 17.7698 13.3018 17.9611 13.1403 18.205C12.9788 18.4489 12.8921 18.7347 12.8909 19.0273V19.1818C12.8909 19.664 12.6994 20.1265 12.3584 20.4675C12.0174 20.8084 11.5549 21 11.0727 21C10.5905 21 10.1281 20.8084 9.78708 20.4675C9.4461 20.1265 9.25455 19.664 9.25455 19.1818V19.1C9.24751 18.7991 9.15011 18.5073 8.97501 18.2625C8.79991 18.0176 8.55521 17.8312 8.27273 17.7273C7.99853 17.6063 7.69437 17.5702 7.39947 17.6236C7.10456 17.6771 6.83244 17.8177 6.61818 18.0273L6.56364 18.0818C6.39478 18.2509 6.19425 18.385 5.97353 18.4765C5.7528 18.568 5.51621 18.6151 5.27727 18.6151C5.03834 18.6151 4.80174 18.568 4.58102 18.4765C4.36029 18.385 4.15977 18.2509 3.99091 18.0818C3.82186 17.913 3.68775 17.7124 3.59626 17.4917C3.50476 17.271 3.45766 17.0344 3.45766 16.7955C3.45766 16.5565 3.50476 16.3199 3.59626 16.0992C3.68775 15.8785 3.82186 15.678 3.99091 15.5091L4.04545 15.4545C4.25503 15.2403 4.39562 14.9682 4.4491 14.6733C4.50257 14.3784 4.46647 14.0742 4.34545 13.8C4.23022 13.5311 4.03887 13.3018 3.79497 13.1403C3.55107 12.9788 3.26526 12.8921 2.97273 12.8909H2.81818C2.33597 12.8909 1.87351 12.6994 1.53253 12.3584C1.19156 12.0174 1 11.5549 1 11.0727C1 10.5905 1.19156 10.1281 1.53253 9.78708C1.87351 9.4461 2.33597 9.25455 2.81818 9.25455H2.9C3.2009 9.24751 3.49273 9.15011 3.73754 8.97501C3.98236 8.79991 4.16883 8.55521 4.27273 8.27273C4.39374 7.99853 4.42984 7.69437 4.37637 7.39947C4.3229 7.10456 4.18231 6.83244 3.97273 6.61818L3.91818 6.56364C3.74913 6.39478 3.61503 6.19425 3.52353 5.97353C3.43203 5.7528 3.38493 5.51621 3.38493 5.27727C3.38493 5.03834 3.43203 4.80174 3.52353 4.58102C3.61503 4.36029 3.74913 4.15977 3.91818 3.99091C4.08704 3.82186 4.28757 3.68775 4.50829 3.59626C4.72901 3.50476 4.96561 3.45766 5.20455 3.45766C5.44348 3.45766 5.68008 3.50476 5.9008 3.59626C6.12152 3.68775 6.32205 3.82186 6.49091 3.99091L6.54545 4.04545C6.75971 4.25503 7.03183 4.39562 7.32674 4.4491C7.62164 4.50257 7.9258 4.46647 8.2 4.34545H8.27273C8.54161 4.23022 8.77093 4.03887 8.93245 3.79497C9.09397 3.55107 9.18065 3.26526 9.18182 2.97273V2.81818C9.18182 2.33597 9.37338 1.87351 9.71435 1.53253C10.0553 1.19156 10.5178 1 11 1C11.4822 1 11.9447 1.19156 12.2856 1.53253C12.6266 1.87351 12.8182 2.33597 12.8182 2.81818V2.9C12.8193 3.19253 12.906 3.47834 13.0676 3.72224C13.2291 3.96614 13.4584 4.15749 13.7273 4.27273C14.0015 4.39374 14.3056 4.42984 14.6005 4.37637C14.8954 4.3229 15.1676 4.18231 15.3818 3.97273L15.4364 3.91818C15.6052 3.74913 15.8057 3.61503 16.0265 3.52353C16.2472 3.43203 16.4838 3.38493 16.7227 3.38493C16.9617 3.38493 17.1983 3.43203 17.419 3.52353C17.6397 3.61503 17.8402 3.74913 18.0091 3.91818C18.1781 4.08704 18.3122 4.28757 18.4037 4.50829C18.4952 4.72901 18.5423 4.96561 18.5423 5.20455C18.5423 5.44348 18.4952 5.68008 18.4037 5.9008C18.3122 6.12152 18.1781 6.32205 18.0091 6.49091L17.9545 6.54545C17.745 6.75971 17.6044 7.03183 17.5509 7.32674C17.4974 7.62164 17.5335 7.9258 17.6545 8.2V8.27273C17.7698 8.54161 17.9611 8.77093 18.205 8.93245C18.4489 9.09397 18.7347 9.18065 19.0273 9.18182H19.1818C19.664 9.18182 20.1265 9.37338 20.4675 9.71435C20.8084 10.0553 21 10.5178 21 11C21 11.4822 20.8084 11.9447 20.4675 12.2856C20.1265 12.6266 19.664 12.8182 19.1818 12.8182H19.1C18.8075 12.8193 18.5217 12.906 18.2778 13.0676C18.0339 13.2291 17.8425 13.4584 17.7273 13.7273Z"
+                                    stroke="currentColor"
+                                    strokeWidth="2"
+                                    strokeLinecap="round"
+                                    strokeLinejoin="round"
+                                ></path>
+                            </svg>
+                        </span>
+                        <span className="block sm:hidden sm:group-hover:inline lg:inline">
+                            Settings
+                        </span>
+                    </button>
+                </li>
+            </ul>
+        </aside>
     );
 }
-
-Sidebar.propTypes = {
-    className: PropTypes.string
-};
 
 export default Sidebar;
