+++ client/src/components/WatchLaterVideos/WatchLaterVideos.jsx
@@ -96,11 +96,6 @@ function WatchLaterVideos() {
                                               )}
                                           </span>
                                       </div>
-                                      {/* Description snippet */}
-                                      <p className="text-sm text-muted-foreground line-clamp-2">
-                                          {watchLaterVideo?.description ||
-                                              "No description available."}
-                                      </p>
                                   </div>
                               </CardContent>
                           </Card>
